backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
chromadb==0.4.18
sentence-transformers==2.2.2
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
pypdf2==3.0.1
python-docx==1.1.0
apscheduler==3.10.4